Change methods' names + change formatting to from 4/8 to 2/4 (indentations).
