+++ build.gradle
@@ -25,7 +25,7 @@ task writePom {
                     license {
                         name 'The Apache Software License, Version 2.0'
                         url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-                        distribution 'com.peacemakr'
+                        distribution 'io.peacemakr'
                     }
                 }
             }
