+++ pom.xml
@@ -10,7 +10,7 @@
     <license>
       <name>The Apache Software License, Version 2.0</name>
       <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
-      <distribution>com.peacemakr</distribution>
+      <distribution>io.peacemakr</distribution>
     </license>
   </licenses>
   <dependencies>
